requirements.txt
# Core dependencies from existing modules
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
anthropic>=0.7.0
openai>=1.0.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0

# Web framework dependencies
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Authentication & Database dependencies
flask-jwt-extended>=4.5.0
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0
flask-migrate>=4.0.0
bcrypt>=4.1.0
email-validator>=2.1.0
cryptography>=41.0.0

# Additional utilities for web deployment
gunicorn>=21.2.0
python-multipart>=0.0.6
psutil>=5.9.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0 